Convert routes to TS
