+++ package.json
@@ -51,8 +51,10 @@
     "typescript": "^5.1.6"
   },
   "dependencies": {
-    "lodash": "^4.17.21",
+    "@types/lodash": "^4.14.198",
+    "@types/qs": "^6.9.8",
     "date-fns": "^2.30.0",
+    "lodash": "^4.17.21",
     "qs": "^6.11.2"
   },
   "peerDependencies": {
